Make models' ids Long instead of Int

+++ app/src/main/java/com/github/amusinamaria/repository/data/Genre.kt
@@ -6,4 +6,4 @@ import kotlinx.serialization.Serializable
 
 @Parcelize
 @Serializable
-data class Genre(val id: Int, val name: String) : Parcelable
+data class Genre(val id: Long, val name: String) : Parcelable
